+++ flexmeasures/utils/unit_utils.py
@@ -1,8 +1,9 @@
 from datetime import timedelta
-from typing import Optional
+from typing import List, Optional, Union
 
 from moneyed import list_all_currencies
 import importlib.resources as pkg_resources
+import pandas as pd
 import pint
 
 # Edit constants template to stop using h to represent planck_constant
@@ -77,9 +78,7 @@ def determine_unit_conversion_multiplier(
     """Determine the value multiplier for a given unit conversion.
     If needed, requires a duration to convert from units of stock change to units of flow.
     """
-    scalar = (
-        ur.Quantity(from_unit).to_base_units() / ur.Quantity(to_unit).to_base_units()
-    )
+    scalar = ur.Quantity(from_unit) / ur.Quantity(to_unit)
     if scalar.dimensionality == ur.Quantity("h").dimensionality:
         if duration is None:
             raise ValueError(
@@ -151,3 +150,36 @@ def is_energy_unit(unit: str) -> bool:
     if not is_valid_unit(unit):
         return False
     return ur.Quantity(unit).dimensionality == ur.Quantity("Wh").dimensionality
+
+
+def convert_units(
+    data: Union[pd.Series, List[Union[int, float]]],
+    from_unit: str,
+    to_unit: str,
+    event_resolution: Optional[timedelta],
+) -> Union[pd.Series, List[Union[int, float]]]:
+    """Updates data values to reflect the given unit conversion."""
+
+    if from_unit != to_unit:
+        try:
+            if isinstance(data, pd.Series):
+                data = pd.Series(
+                    pint.Quantity(data.values, from_unit)
+                    .to(pint.Quantity(to_unit))
+                    .magnitude,
+                    index=data.index,
+                    name=data.name,
+                )
+            else:
+                data = list(
+                    pint.Quantity(data, from_unit).to(pint.Quantity(to_unit)).magnitude
+                )
+        except pint.errors.DimensionalityError:
+            multiplier = determine_unit_conversion_multiplier(
+                from_unit, to_unit, event_resolution
+            )
+            if isinstance(data, pd.Series):
+                data = multiplier * data
+            else:
+                data = [multiplier * value for value in data]
+    return data